make new messages library and fix udp

+++ Shared/Gui/Gui/mainwindow.cpp
@@ -36,7 +36,7 @@ void MainWindow::addConnection(QHostAddress ip, quint16 port)
 void MainWindow::receivedMsg(std::string s, QHostAddress ip, quint16 port)
 {
   auto t = ip.toString();
-  if (t.contains("::ffff:"))t = t.mid(7, s.size() - 7);
+  if (t.contains("::ffff:"))t = t.mid(7, static_cast<int>(s.size()) - 7);
   QString toOutput = t + ": " + QString::number(port) + "\n" + QString::fromStdString(s);
   ui->label_received->setText(toOutput);
 }
